AMBARI-793. Make MapReduce, Nagios, and Ganglia optional during cluster install. (yusaku)
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/branches/AMBARI-666@1393286 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-web/app/controllers/main/admin/user.js
@@ -19,6 +19,12 @@
 App.MainAdminUserController = Em.Controller.extend({
   name:'mainAdminUserController',
   deleteRecord:function (event) {
+
+    if(event.context.get('userName') == App.get('router').getLoginName()) {
+      alert(Em.I18n.t('admin.users.deleteYourselfMessage'));
+      return ;
+    }
+
     if (Em.I18n.t('question.sure')) {
       event.context.deleteRecord();
       App.store.commit();
